+++ app/Http/Controllers/LoginControllers/AdminLoginController.php
@@ -37,6 +37,8 @@ class AdminLoginController extends Controller
         $company->strCompanyName = $request->input('companyname');
         $company->strCompanyAddress = $request->input('address');
         $company->strCompanyEmail = $request->input('email');
+        $company->strCompanyContactPNum = $request->input('mobilenum');
+        $company->strCompanyContactTNum = $request->input('telnum');
         $company->save();
         // return response(['username'=>$request->input('username')]);
         $user = User::create([